+++ packages/atomic/src/components/commerce/atomic-commerce-search-box/atomic-commerce-search-box.ts
@@ -30,9 +30,9 @@ import {
 import {updateBreakpoints} from '@/src/utils/replace-breakpoint-utils';
 import {getDefaultSlotContent} from '@/src/utils/slot-utils';
 import {isFocusingOut} from '../../../utils/dom-utils';
+import {once} from '../../../utils/function-utils';
 import {spreadProperties} from '../../../utils/object-utils';
 import {randomID} from '../../../utils/random-utils';
-import {once} from '../../../utils/utils';
 import type {RedirectionPayload} from '../../common/search-box/redirection-payload';
 import {renderSearchBoxWrapper} from '../../common/search-box/search-box-wrapper';
 import {renderSearchBoxTextArea} from '../../common/search-box/search-text-area';
