+++ packages/atomic/src/components/common/generated-answer/atomic-generated-answer-feedback/atomic-generated-answer-feedback-modal.tsx
@@ -20,7 +20,7 @@ import {
   InitializeBindings,
 } from '../../../../utils/initialization-utils';
 import {updateBreakpoints} from '../../../../utils/replace-breakpoint-utils';
-import {once} from '../../../../utils/utils';
+import {once} from '../../../../utils/function-utils';
 import {randomID} from '../../../../utils/random-utils';
 import {ATOMIC_MODAL_EXPORT_PARTS} from '../../atomic-modal/export-parts';
 import {IconButton} from '../../stencil-iconButton';
