+++ packages/atomic/src/components/search/atomic-search-box-recent-queries/atomic-search-box-recent-queries.ts
@@ -7,8 +7,8 @@ import {html, LitElement, nothing} from 'lit';
 import {customElement, property, state} from 'lit/decorators.js';
 import {errorGuard} from '@/src/decorators/error-guard';
 import type {SearchBoxSuggestionsComponent} from '@/src/decorators/types';
+import {once} from '@/src/utils/function-utils';
 import {SafeStorage, StorageItems} from '@/src/utils/local-storage-utils';
-import {once} from '@/src/utils/utils';
 import Clock from '../../../images/clock.svg';
 import {
   getPartialRecentQueryClearElement,
